requirements.txt
streamlit==1.29.0
pymongo==4.6.0
ccxt==4.1.66
bcrypt==4.1.2
cryptography==41.0.8
python-dotenv==1.0.0
pandas==2.1.4